requirements.txt
websockets==12.0
opencv-python>=4.8
Pillow>=10.0.0
pyyaml>=6.0
numpy<2.3,>=2.0
llama-cpp-python>=0.3.16
sounddevice>=0.5.3
requests>=2.31.0
diskcache>=5.6.1
Jinja2>=3.1.4
webrtcvad==2.0.10
vosk>=0.3.45
soundfile>=0.12.1
setuptools